[update] move implementation of byte_histogram to declaration
